Refactor config object
