--- jest/jest.setup.js
@@ -1,2 +1,0 @@
-// globalize config made available via webpack provide plugin
-window.config = require('../config.js');
