+++ src/components/AutocompleteGeocoder/apiService.js
@@ -1,9 +1,10 @@
 import logger from '~/utils/logger';
+import mapConfig from '~/pages/Map/config/map';
 
 let abortController = new window.AbortController();
 
 function compileSearchUrl(searchString) {
-  const { accessToken, geocoderBounds } = config.map;
+  const { accessToken, geocoderBounds } = mapConfig;
   return (
     `https://api.mapbox.com/geocoding/v5/mapbox.places/${searchString}.json?` +
     `access_token=${accessToken}&autocomplete=true&language=de&` +
