+++ src/config/colors.ts
@@ -1,0 +1,16 @@
+export default {
+  lightbg: '#f5f5f5',
+  lightgrey: '#e8e8e8',
+  inactivegrey: '#cccccc',
+  midgrey: '#9b9b9b',
+  darkgrey: '#545454',
+  darkbg: '#353535',
+  black: '#0f0f0f',
+  white: '#FFF',
+  interaction: '#fabe28',
+  change_4: '#910055',
+  index: '#ff650c',
+  error: '#FF5050',
+  likebg: '#FFF9EC',
+  katasterHighlight: '#fabe28'
+};
