+++ src/config/staticpages.ts
@@ -1,0 +1,38 @@
+export default [
+  {
+    key: 'about',
+    route: '/info'
+  },
+  {
+    key: 'contact',
+    route: '/kontakt'
+  },
+  {
+    key: 'privacy',
+    route: '/datenschutz'
+  },
+  {
+    key: 'imprint',
+    route: '/impressum'
+  },
+  {
+    key: 'press',
+    route: '/presse'
+  },
+  {
+    key: 'faq',
+    route: '/faq'
+  },
+  {
+    key: 'api',
+    route: '/api'
+  },
+  {
+    key: 'jobs',
+    route: '/jobs'
+  },
+  {
+    key: 'newsletter-thank-you',
+    route: '/newsletter-danke'
+  }
+];
