+++ src/services/reverseGeocode.js
@@ -1,4 +1,5 @@
 import ky from 'ky';
+import config from '~/pages/Map/config';
 
 export default async ({ lat, lng }) => {
   // form a url following the docs https://docs.mapbox.com/api/search/#reverse-geocoding
