+++ typescript/custom.d.ts
@@ -2,5 +2,3 @@ declare module '*.svg' {
   const content: any;
   export default content;
 }
-
-declare const config: any;
